pyproject.toml
[project]
name = "text-to-sql-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "flask>=3.1.0",
    "openai>=1.56.1",
    "polars>=1.17.1",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.3",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "simplejson>=3.19.3",
    "sql-metadata>=2.15.0",
    "sqlglot>=26.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.1",
]

[tool.ruff.lint]
extend-select = ["B", "I"]
